+++ external/vcm/vcm/calc/thermo/constants.py
